added documentation and cleaned up the code
