+++ src/components/HomePage.jsx
@@ -7,8 +7,6 @@ class HomePage extends Component {
   render(){
     return(
       <div className="home">
-
-
         <div className="header">
           MARS
           <div className="subtitle">
