--- src/components/helpers/mapSamples.js
@@ -1,3 +1,0 @@
-//map the user's samples to SESAR's specs according to the uploaded mapping file
-//inputs: sample csvs and mapping file
-//output: json file/object that is correctly mapped
