+++ src/index.js
@@ -15,34 +15,23 @@ import HomePage from './components/HomePage';
 import SignOut from './components/SignOut';
 import Details from './components/Details'
 import Header from './components/NavBar';
-
-//CSS
-//import 'bootstrap/dist/css/bootstrap.css';
-
-//Actions and Reducers
-import reducers from './reducers';
 import { loadState, saveState } from './localstorage';
 
+//Reducer
+import reducers from './reducers';
+
+//create store
 const persistedState = loadState();
 const createStoreWithMiddleware = applyMiddleware(reduxThunk)(createStore);
 const store = createStoreWithMiddleware(reducers, persistedState,
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__())
 
-//Check cookies to see if user is already authenticated
-//and login user is already authenticated
-
+//Check persisted state for any authentication data
 store.subscribe(() => {
   saveState({
     auth: store.getState().auth
   });
 });
-/*console.log('Cookies usercode:', usercode)
-
-if(usercode) {
-  store.dispatch({ type: AUTHENTICATED });
-}*/
-
-
 
 ReactDOM.render(
   <Provider store={store}>
