+++ src/reducers/upload_reducer.js
@@ -1,12 +1,29 @@
-import {CHANGE_SOURCE_FILE, CHANGE_MAP_FILE} from '../actions'
+import {UPLOAD_REQUEST, UPLOAD_SUCCESS, UPLOAD_FAILURE, INITIALIZE_SAMPLES} from '../actions'
 
-export default function(state={}, action){
-    switch (action.type){
-        case CHANGE_SOURCE_FILE:
-            return {...state, sourceFiles: action.sourceFiles}
-        case CHANGE_MAP_FILE:
-            return {...state, mapFile: action.mapFile}
-        default:
-            return state;   
-    }
+export default function reducer(state = {}, action) {
+  switch(action.type) {
+    case INITIALIZE_SAMPLES:
+        console.log("<==== Samples Ready ====>")
+        return {...state, samples: action.sampleArray}
+    case UPLOAD_REQUEST:
+        console.log("<==== Upload Requested ====>")
+      return {...state, loading: true}
+    case UPLOAD_SUCCESS:
+      let results = action.results
+      let samples = state.samples
+
+      //Add the IGSNs to each sample
+      for(let i=0; i<results.length; i++) {
+        let igsn = {originalKey: '', originalValue: '', key:'igsn', value:results[i].igsn}  //IGSN for each sample
+        samples[i] = [...samples[i], igsn]  //for each sample, the sample is equal to its previous version with IGSN added to the end
+      }
+      //TODO: enable exporting to CSV after successful upload
+      console.log("<==== Upload Succcessful ====>")
+      return {...state, samples: samples, loading: false}
+    case UPLOAD_FAILURE:
+      console.log("<==== Upload Failure ====>")
+      return {...state, loading: false}
+    default:
+      return state
+  }
 }
